+++ app/src/main/java/com/example/shinhan_qna_aos/login/LoginScreen.kt
@@ -23,180 +23,17 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.shinhan_qna_aos.BuildConfig
 import com.example.shinhan_qna_aos.R
-import com.example.shinhan_qna_aos.sendGoogleAuthCodeToServer
-import com.example.shinhan_qna_aos.sendKakaoOpenIdTokenToServer
+import com.example.shinhan_qna_aos.login.loginserver.sendGoogleAuthCodeToServer
+import com.example.shinhan_qna_aos.login.loginserver.sendKakaoOpenIdTokenToServer
 import com.example.shinhan_qna_aos.ui.theme.pretendard
 import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.kakao.sdk.auth.AuthApiClient
 import com.kakao.sdk.user.UserApiClient
-
-//@Composable
-//fun LoginScreen() {
-//    val context = LocalContext.current
-//    val TAG = "kakao"
-//    // OpenID Connect를 위한 스코프 추가
-//    val kakaoAuthUrl =
-//        "https://kauth.kakao.com/oauth/authorize" +
-//                "?client_id=${BuildConfig.KAKAO_REST_API}" + // REST API 키
-//                "&redirect_uri=${BuildConfig.LOCAL_URL}/oauth/callback/kakao" + // 백엔드 서버 주소
-//                "&response_type=code"
-//
-//    // 카카오 로그인 시도 런처
-//    val kakaoAuthLauncher = rememberLauncherForActivityResult(
-//        contract = ActivityResultContracts.StartActivityForResult()
-//    ) { result ->
-//        // 로그인 결과 처리
-//        if (result.resultCode == Activity.RESULT_OK) {
-//            val tokenManager = AuthApiClient.instance.tokenManager
-//            val tokenInfo = tokenManager.getToken()
-//
-//            tokenInfo?.let {
-//                val accessToken = it.accessToken
-//                val idToken = it.idToken // OpenID 토큰
-//
-//                if (idToken != null) {
-//                    // TODO: OpenID 토큰을 백엔드 서버로 전송하는 함수 호출
-//                    sendKakaoOpenIdTokenToServer(idToken, TAG)
-//                } else {
-//                    Log.d(TAG, "ID Token is null. OpenID Connect is required.")
-//                }
-//            } ?: run {
-//                Log.e(TAG, "카카오 토큰 정보 가져오기 실패")
-//            }
-//        } else {
-//            Log.e(TAG, "카카오 로그인 실패: ${result.data?.getStringExtra("error")}")
-//        }
-//    }
-//
-//    val googleSignInClient = getGoogleSignInClient(context)
-//
-//    // Google 로그인 결과를 처리하는 런처
-//    val googleAuthLauncher = rememberLauncherForActivityResult(
-//        contract = ActivityResultContracts.StartActivityForResult()
-//    ) { result ->
-//        val task = GoogleSignIn.getSignedInAccountFromIntent(result.data)
-//        try {
-//            val account = task.getResult(com.google.android.gms.common.api.ApiException::class.java)
-//            val authCode = account.serverAuthCode
-//            Log.d("GoogleLogin", "받은 인가코드: $authCode")
-//
-//            if (authCode != null) {
-//                sendGoogleAuthCodeToServer(authCode,"google")
-//            }
-//        } catch (e: com.google.android.gms.common.api.ApiException) {
-//            Log.e("GoogleLogin", "Google 로그인 실패: ${e.statusCode}")
-//        }
-//    }
-//
-//    BoxWithConstraints(
-//        modifier = Modifier
-//            .fillMaxSize()
-//            .systemBarsPadding()
-//            .background(Color.White)
-//            .padding(top = 64.dp, start = 40.dp, end = 40.dp, bottom = 48.dp),
-//        contentAlignment = Alignment.Center
-//    ) {
-//        val maxwidth=maxWidth
-//        Column(
-//            modifier = Modifier.fillMaxSize(),
-//            horizontalAlignment = Alignment.CenterHorizontally,
-//            verticalArrangement = Arrangement.SpaceBetween
-//        ) {
-//            Image(
-//                painter = painterResource(R.drawable.biglogo),
-//                contentDescription = null,
-//                modifier = Modifier.size(128.dp)
-//            )
-//
-//            Column(
-//                modifier = Modifier.fillMaxWidth(),
-//                horizontalAlignment = Alignment.CenterHorizontally
-//            ) {
-//                // 카카오 로그인 버튼
-//                Image(
-//                    painter = painterResource(R.drawable.kakao_login),
-//                    contentDescription = "카카오 로그인",
-//                    modifier = Modifier
-//                        .fillMaxWidth()
-//                        .clickable {
-//                            if (UserApiClient.instance.isKakaoTalkLoginAvailable(context)) {
-//                                UserApiClient.instance.loginWithKakaoTalk(context) { token, error ->
-//                                    if (error != null) {
-//                                        Log.e(TAG, "카카오톡 로그인 실패", error)
-//                                        // 카카오 계정으로 로그인 시도
-//                                        UserApiClient.instance.loginWithKakaoAccount(context) { token, error ->
-//                                            if (error != null) {
-//                                                Log.e(TAG, "카카오 계정 로그인 실패", error)
-//                                            } else if (token != null) {
-//                                                Log.d(TAG, "카카오 계정 로그인 성공")
-//                                                // TODO: 토큰 정보를 백엔드 서버로 전송
-//                                                token.idToken?.let { idToken ->
-//                                                    sendKakaoOpenIdTokenToServer(idToken, TAG)
-//                                                }
-//                                            }
-//                                        }
-//                                    } else if (token != null) {
-//                                        Log.d(TAG, "카카오톡 로그인 성공")
-//                                        // TODO: 토큰 정보를 백엔드 서버로 전송
-//                                        token.idToken?.let { idToken ->
-//                                            sendKakaoOpenIdTokenToServer(idToken, TAG)
-//                                        }
-//                                    }
-//                                }
-//                            } else {
-//                                UserApiClient.instance.loginWithKakaoAccount(context) { token, error ->
-//                                    if (error != null) {
-//                                        Log.e(TAG, "카카오 계정 로그인 실패", error)
-//                                    } else if (token != null) {
-//                                        Log.d(TAG, "카카오 계정 로그인 성공")
-//                                        // TODO: 토큰 정보를 백엔드 서버로 전송
-//                                        token.idToken?.let { idToken ->
-//                                            sendKakaoOpenIdTokenToServer(idToken, TAG)
-//                                        }
-//                                    }
-//                                }
-//                            }
-//                        }
-//                )
-//                Spacer(modifier = Modifier.height(12.dp))
-//
-//                // 구글 로그인 버튼
-//                Image(
-//                    painter = painterResource(R.drawable.google_login),
-//                    contentDescription = "구글 로그인",
-//                    modifier = Modifier
-//                        .fillMaxWidth()
-//                        .clickable {
-//                            googleAuthLauncher.launch(googleSignInClient.signInIntent)
-//                        }
-//                )
-//                Spacer(modifier = Modifier.height(12.dp))
-//
-//                Text(
-//                    text = "관리자 전용 페이지",
-//                    color = Color(0xffDFDFDF),
-//                    style = TextStyle(
-//                        fontFamily = pretendard,
-//                        fontWeight = FontWeight.Normal,
-//                        fontSize = 12.sp
-//                    ),
-//                    modifier = Modifier.clickable { /* 관리자 로그인 처리 */ }
-//                )
-//            }
-//        }
-//    }
-//}
 
 @Composable
 fun LoginScreen() {
     val context = LocalContext.current
     val TAG = "kakao"
-
-    // 카카오 SDK를 사용하는 경우, 이 런처는 필요 없으므로 제거합니다.
-    // val kakaoAuthUrl = "..."
-    // val kakaoAuthLauncher = rememberLauncherForActivityResult(...)
 
     val googleSignInClient = getGoogleSignInClient(context)
